fix: factor out utils
