+++ packit_service_fedmsg/consumer.py
@@ -2,15 +2,15 @@
 # SPDX-License-Identifier: MIT
 
 from datetime import datetime, timezone
-from functools import reduce
 from logging import getLogger
 from os import getenv
 from pathlib import Path
-from typing import Any
 
 from celery import Celery
 from fedora_messaging import config
 from fedora_messaging.message import Message
+
+from packit_service_fedmsg.utils import nested_get, specfile_changed
 
 config.conf.setup_logging()
 logger = getLogger(__name__)
@@ -38,52 +38,6 @@ DISTGIT_PR_FLAG_TOPIC = {
 NEW_HOTNESS_TOPIC = "org.fedoraproject.prod.hotness.update.bug.file"
 
 DISTGIT_PR_COMMENT_ADDED = "org.fedoraproject.prod.pagure.pull-request.comment.added"
-
-
-def nested_get(d: dict, *keys, default=None) -> Any:
-    """
-    Recursively obtain value from nested dict.
-
-    Args:
-        d: Dictionary to get value from.
-        keys: Path within the dictionary.
-        default: Value to be returned if some key is not found.
-
-            Defaults to `None`.
-
-    Returns:
-        Value found in the dictionary or specified default.
-    """
-    response = d
-
-    try:
-        for k in keys:
-            response = response[k]
-    except (KeyError, AttributeError, TypeError):
-        # logger.debug("can't obtain %s: %s", k, ex)
-        return default
-
-    return response
-
-
-def specfile_changed(body: dict) -> bool:
-    """
-    Does the commit contain specfile change?
-
-    Args:
-        body: Body of the message.
-
-    Returns:
-        `True` if the specfile has changed, `False` otherwise.
-    """
-    files = reduce(
-        lambda val, key: val.get(key) if val else None,
-        ["commit", "stats", "files"],
-        body,
-    )
-    file_names = files.keys() if files else []
-
-    return any(file_name.endswith(".spec") for file_name in file_names)
 
 
 class Consumerino:
