+++ tests/test_utils.py
@@ -3,7 +3,7 @@
 
 import pytest
 
-from packit_service_fedmsg.consumer import specfile_changed
+from packit_service_fedmsg.utils import specfile_changed
 
 
 @pytest.mark.parametrize(
